Let the frontened task communicate with worker task to complete the calculation job.
Everytime the test command receive the calculation job, it send the necessary data to the queue! Another task(worker task) receive the data and starting calculation,when finished its job, it put the result to another queue.In main.c file, a function will receive the result from the queue and print the result.This make the typesetting look better.I also add an function to return the queue handle so that in the main.c file can also get the queue handle.